refactor plugin
