+++ plugins/quetz_runexports/setup.py
@@ -1,0 +1,8 @@
+from setuptools import setup
+
+setup(
+    name="quetz-runexports",
+    install_requires="quetz",
+    entry_points={"quetz": ["quetz-runexports = quetz_runexports.main"]},
+    package=["quetz_runexports"],
+)
