--- sample_plugin/setup.py
@@ -1,8 +1,0 @@
-from setuptools import setup
-
-setup(
-    name="quetz-runexports",
-    install_requires="quetz",
-    entry_points={"quetz": ["quetz-runexports = main"]},
-    py_modules=["main"],
-)
